requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
aioodbc==0.5.0
pyodbc==5.3.0
greenlet==3.0.3
pydantic==2.6.3
pydantic-settings==2.2.1
jinja2==3.1.3
python-multipart==0.0.9
apscheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.1.0
pytest-asyncio==0.23.5
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
itsdangerous==2.2.0
starlette==0.36.3
aiosqlite
bcrypt==4.0.1
